fix header cutoff
